project2/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
tensorflow==2.20.0
numpy==2.0.2
Pillow==11.3.0
python-multipart==0.0.20